refactor: 코스 디테일 조회 api 리얼타임 호출으로 수정(#82)
