+++ src/main/java/com/joyride/ms/src/course/CourseDao.java
@@ -54,31 +54,31 @@ public class CourseDao {
         this.jdbcTemplate.update(createCourseListQuery, createCourseListParams);
     }
 
-    // 코스 디테일 정보
-    public GetCourseRes selectCourse(String course_id){
-        String getCourseQuery = "select id,title,course_img_url,content,summary," +
-                "tour_point, travelerinfo, distance, difficulty, sigun, required_at, created_at, updated_at " +
-                "from course where id = ?";
-
-        String getCourseParams = course_id;
-        //db정보 가져오기
-        return this.jdbcTemplate.queryForObject(getCourseQuery,
-                (rs,rowNum) -> GetCourseRes.createGetCourseRes(
-                        rs.getInt("id"),
-                        rs.getString("title"),
-                        rs.getString("course_img_url"),
-                        rs.getString("content"),
-                        rs.getString("summary"),
-                        rs.getString("tour_point"),
-                        rs.getString("travelerinfo"),
-                        rs.getDouble("distance"),
-                        rs.getInt("difficulty"),
-                        rs.getString("sigun"),
-                        rs.getDouble("required_at"),
-                        rs.getString("created_at"),
-                        rs.getString("updated_at")
-                ), getCourseParams);
-    }
+//    // 코스 디테일 정보
+//    public GetCourseRes selectCourse(String course_id){
+//        String getCourseQuery = "select id,title,course_img_url,content,summary," +
+//                "tour_point, travelerinfo, distance, difficulty, sigun, required_at, created_at, updated_at " +
+//                "from course where id = ?";
+//
+//        String getCourseParams = course_id;
+//        //db정보 가져오기
+//        return this.jdbcTemplate.queryForObject(getCourseQuery,
+//                (rs,rowNum) -> GetCourseRes.createGetCourseRes(
+//                        rs.getInt("id"),
+//                        rs.getString("title"),
+//                        rs.getString("course_img_url"),
+//                        rs.getString("content"),
+//                        rs.getString("summary"),
+//                        rs.getString("tour_point"),
+//                        rs.getString("travelerinfo"),
+//                        rs.getDouble("distance"),
+//                        rs.getInt("difficulty"),
+//                        rs.getString("sigun"),
+//                        rs.getDouble("required_at"),
+//                        rs.getString("created_at"),
+//                        rs.getString("updated_at")
+//                ), getCourseParams);
+//    }
 
     //코스 좋아요 유저 아이디 조회
     public List<Integer> selectUserIdByCourseId(String course_id) {
