+++ src/main/java/com/joyride/ms/src/course/CourseService.java
@@ -22,148 +22,8 @@ import static com.joyride.ms.util.BaseResponseStatus.*;
 @Transactional
 @RequiredArgsConstructor
 public class CourseService {
-
-    @Value("${durunubi.secret}")
-    private String API_SECRET_KEY;
     private final CourseDao courseDao;
     private final CourseProvider courseProvider;
-
-    // courseArr를 리턴해주는 method
-    public JSONArray callCourseAPI() throws Exception {
-        try {
-            String result = "";
-
-            URL url = new URL("https://api.visitkorea.or.kr/openapi/service/rest/Durunubi/" +
-                    "courseList?MobileOS=ETC&MobileApp=joyride&ServiceKey=" + API_SECRET_KEY +
-                    "&brdDiv=DNBW&numOfRows=3004&pageNo=1&_type=json");
-
-            BufferedReader bf;
-            bf = new BufferedReader(new InputStreamReader(url.openStream(), "UTF-8"));
-            result = bf.readLine();
-
-            JSONParser jsonParser = new JSONParser();
-            JSONObject jsonObject = (JSONObject) jsonParser.parse(result);
-            JSONObject response = (JSONObject) jsonObject.get("response");
-            JSONObject body = (JSONObject) response.get("body");
-            JSONObject items = (JSONObject) body.get("items");
-            JSONArray courseArr = (JSONArray) items.get("item");
-            return courseArr;
-        }
-        catch (Exception exception) {
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
-    // 디테일 조회시 api 호출(String title을 파라미터로 받는)
-    public JSONArray callCourseAPI(String title) throws Exception {
-        try {
-            String result = "";
-
-            URL url = new URL("https://api.visitkorea.or.kr/openapi/service/rest/Durunubi/" +
-                    "courseList?MobileOS=ETC&MobileApp=joyride&ServiceKey=" + API_SECRET_KEY +
-                    "&crsKorNm="+title+"&_type=json");
-
-            BufferedReader bf;
-            bf = new BufferedReader(new InputStreamReader(url.openStream(), "UTF-8"));
-            result = bf.readLine();
-
-            JSONParser jsonParser = new JSONParser();
-            JSONObject jsonObject = (JSONObject) jsonParser.parse(result);
-            JSONObject response = (JSONObject) jsonObject.get("response");
-            JSONObject body = (JSONObject) response.get("body");
-            JSONObject items = (JSONObject) body.get("items");
-            JSONArray courseArr = (JSONArray) items.get("item");
-            return courseArr;
-        }
-        catch (Exception exception) {
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
-
-
-    // 반환 리스트를 만들어주는 메소드
-    public List<GetCourseListRes> createCourseList(JSONArray courseArr) throws BaseException {
-        try {
-            // 중복 확인용 리스트
-            ArrayList<String> courseIdList = new ArrayList<>();
-            // 반환할 리스트 객체 생성
-            ArrayList<GetCourseListRes> getCourseList = new ArrayList<>();
-
-            for (int i = 0; i < courseArr.size(); i++) {
-                JSONObject course = (JSONObject)courseArr.get(i);
-                //중복 제거 및, 자전거 길만
-                if (courseIdList.contains((String)course.get("crsIdx"))) {
-                    continue;
-                } else {
-                    courseIdList.add((String)course.get("crsIdx"));
-                }
-
-                String checkBrdDiv = (String)course.get("brdDiv");
-                if (checkBrdDiv.equals("DNWW")) {
-                    continue;
-                }
-                String crsIdx = (String)course.get("crsIdx");
-                String crsKorNm = (String)course.get("crsKorNm");
-                String crsContents = (String)course.get("crsContents");
-                String crsSummary = (String)course.get("crsSummary");
-                String crsTourInfo = (String)course.get("crsTourInfo");
-                String travelerinfo = (String)course.get("travelerinfo");
-                String crsDstncStr = (String)course.get("crsDstnc");
-                double crsDstnc = Double.parseDouble(crsDstncStr);
-                String crsLevelStr = (String)course.get("crsLevel");
-                int crsLevel = Integer.parseInt(crsLevelStr);
-                String sigun = (String)course.get("sigun");
-                String crsTotlRqrmHourStr = (String)course.get("crsTotlRqrmHour");
-                double crsTotlRqrmHour = Double.parseDouble(crsTotlRqrmHourStr);
-
-                // 일단 이미지는 없다고 생각하고.
-                GetCourseListRes getCourseListRes = GetCourseListRes.createGetCourseListRes(crsIdx, crsKorNm, crsContents, crsSummary,
-                        crsTourInfo, travelerinfo, crsDstnc, crsLevel, sigun, crsTotlRqrmHour);
-
-                String courseId = getCourseListRes.getId();
-                int likeCount = courseProvider.retrieveCourseLikeCount(courseId);
-                getCourseListRes.setLikeCount(likeCount);
-
-                getCourseList.add(getCourseListRes);
-
-            }
-            return getCourseList;
-        }
-        catch (Exception exception) {
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
-    // 반환 디테일 코스를 만들어 주는 메소드
-    public GetCourseRes createCourse(JSONArray courseArr) throws BaseException {
-        try {
-            JSONObject course = (JSONObject)courseArr.get(0);
-
-            String crsIdx = (String)course.get("crsIdx");
-            String crsKorNm = (String)course.get("crsKorNm");
-            String crsContents = (String)course.get("crsContents");
-            String crsSummary = (String)course.get("crsSummary");
-            String crsTourInfo = (String)course.get("crsTourInfo");
-            String travelerinfo = (String)course.get("travelerinfo");
-            String crsDstncStr = (String)course.get("crsDstnc");
-            double crsDstnc = Double.parseDouble(crsDstncStr);
-            String crsLevelStr = (String)course.get("crsLevel");
-            int crsLevel = Integer.parseInt(crsLevelStr);
-            String sigun = (String)course.get("sigun");
-            String crsTotlRqrmHourStr = (String)course.get("crsTotlRqrmHour");
-            double crsTotlRqrmHour = Double.parseDouble(crsTotlRqrmHourStr);
-
-            // 일단 이미지는 없다고 생각하고.
-            GetCourseRes getCourseRes = GetCourseRes.createGetCourseRes(crsIdx, crsKorNm, crsContents, crsSummary,
-                            crsTourInfo, travelerinfo, crsDstnc, crsLevel, sigun, crsTotlRqrmHour);
-            return getCourseRes;
-        }
-        catch (Exception exception) {
-            throw new BaseException(DATABASE_ERROR);
-        }
-    }
-
 
     // 리뷰작성 service
     public PostCourseReviewRes createCourseReview(PostCourseReviewReq postCourseReviewReq) throws BaseException {
