+++ src/main/java/com/joyride/ms/src/course/model/GetCourseListRes.java
@@ -1,8 +1,17 @@
 package com.joyride.ms.src.course.model;
 
+import com.joyride.ms.src.course.CourseDao;
+import com.joyride.ms.util.BaseException;
 import lombok.*;
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import javax.persistence.Lob;
+import java.util.ArrayList;
+import java.util.List;
+
+import static com.joyride.ms.util.BaseResponseStatus.DATABASE_ERROR;
 
 @Getter
 @Setter
@@ -74,5 +83,54 @@ public class GetCourseListRes {
 
 
         return getCourseListRes;
+    }
+
+    // 반환 리스트를 만들어주는 메소드
+    public static List<GetCourseListRes> createCourseList(JSONArray courseArr) throws BaseException {
+        try {
+            // 중복 확인용 리스트
+            ArrayList<String> courseIdList = new ArrayList<>();
+            // 반환할 리스트 객체 생성
+            ArrayList<GetCourseListRes> getCourseList = new ArrayList<>();
+
+            for (int i = 0; i < courseArr.size(); i++) {
+                JSONObject course = (JSONObject)courseArr.get(i);
+                //중복 제거 및, 자전거 길만
+                if (courseIdList.contains((String)course.get("crsIdx"))) {
+                    continue;
+                } else {
+                    courseIdList.add((String)course.get("crsIdx"));
+                }
+
+                String checkBrdDiv = (String)course.get("brdDiv");
+                if (checkBrdDiv.equals("DNWW")) {
+                    continue;
+                }
+                String crsIdx = (String)course.get("crsIdx");
+                String crsKorNm = (String)course.get("crsKorNm");
+                String crsContents = (String)course.get("crsContents");
+                String crsSummary = (String)course.get("crsSummary");
+                String crsTourInfo = (String)course.get("crsTourInfo");
+                String travelerinfo = (String)course.get("travelerinfo");
+                String crsDstncStr = (String)course.get("crsDstnc");
+                double crsDstnc = Double.parseDouble(crsDstncStr);
+                String crsLevelStr = (String)course.get("crsLevel");
+                int crsLevel = Integer.parseInt(crsLevelStr);
+                String sigun = (String)course.get("sigun");
+                String crsTotlRqrmHourStr = (String)course.get("crsTotlRqrmHour");
+                double crsTotlRqrmHour = Double.parseDouble(crsTotlRqrmHourStr);
+
+                // 일단 이미지는 없다고 생각하고.
+                GetCourseListRes getCourseListRes = GetCourseListRes.createGetCourseListRes(crsIdx, crsKorNm, crsContents, crsSummary,
+                        crsTourInfo, travelerinfo, crsDstnc, crsLevel, sigun, crsTotlRqrmHour);
+
+                getCourseList.add(getCourseListRes);
+
+            }
+            return getCourseList;
+        }
+        catch (Exception exception) {
+            throw new BaseException(DATABASE_ERROR);
+        }
     }
 }
